Remove non-print characters from $prop_nfshost_ip
This was causing an issue on Windows machines when the carriage-return
was included as part of the variable value. When rendering the value in
the configureBoot2Docker step, it would reset the cursor to the
beginning of the line and overwrite the already-printed data. This
ultimately would lead to a failure to mount the NFS mounts.